Improved profile view and added a change password option
Made profile view get data from the local storage about the user, and change it in the view and in the local storage as well.
Made a change password option which requires the current user password from the local storage and changes it to whatever I want if it passes all the validations.