CreatePost/CreateTheme/UserThemesPage/NoThemes
Made a create post page with which you can add a post to a theme.

Fixed create theme functionality to work correctly.

Implemented a user page which shows all of the user's theme with buttons which reveal all of the theme's posts.

Removed old stuff that needs reworking.

Implemented delete theme functionality in the server api which is currently not working.
Working on a way to fix it!