Fixed delete theme functionality and added more feautures
When the user's themes page gets empty, added a component to display an empty page with a reference to the create a theme component.
Added a functionality:
When the user deletes his/hers last theme, the empty page component shows up and when he/she adds one the your themes list shows.

Changed the home page completely.
Added a dynamic list on right side to display recent posts.
Home page still in progress.