Made changes and improvements to existing feautures
Fixed footer not showing correctly on all pages.

Added validation that ensures two themes with the same name cannot be created.

Added a logo to top to navigate to home.

Fixed delete theme error I had with backend by configuring files in the api.

Changed routes a bit.

Added a delete and edit post functionality in the recent theme posts and in the user one's which let you delete and alter the post you've created.

Added a limit to all themes pages to show max number of 10 per page and next and a previous button to navigate further and backwards.